+++ timeutil/pkg/backoff/constant_test.go
@@ -1,6 +1,7 @@
 package backoff_test
 
 import (
+	"context"
 	"testing"
 	"time"
 
@@ -10,11 +11,13 @@ import (
 )
 
 func TestConstant(t *testing.T) {
+	ctx := context.Background()
+
 	g, err := backoff.Constant(5 * time.Second).New()
 	require.NoError(t, err)
 
 	for i := 0; i < 100; i++ {
-		wait, err := g.Next()
+		wait, err := g.Next(ctx)
 		require.NoError(t, err)
 		assert.Equal(t, 5*time.Second, wait, "iteration #%d", i)
 	}