+++ timeutil/pkg/backoff/linear_test.go
@@ -1,6 +1,7 @@
 package backoff_test
 
 import (
+	"context"
 	"math"
 	"testing"
 	"time"
@@ -11,17 +12,21 @@ import (
 )
 
 func TestLinear(t *testing.T) {
+	ctx := context.Background()
+
 	g, err := backoff.Linear(5 * time.Second).New()
 	require.NoError(t, err)
 
 	for i := 0; i < 100; i++ {
-		wait, err := g.Next()
+		wait, err := g.Next(ctx)
 		require.NoError(t, err)
 		assert.Equal(t, 5*time.Duration(i+1)*time.Second, wait, "iteration #%d", i)
 	}
 }
 
 func TestLinearWraparound(t *testing.T) {
+	ctx := context.Background()
+
 	g, err := backoff.Linear(1 << 61).New()
 	require.NoError(t, err)
 
@@ -34,7 +39,7 @@ func TestLinearWraparound(t *testing.T) {
 		math.MaxInt64,
 	}
 	for i, step := range expected {
-		wait, err := g.Next()
+		wait, err := g.Next(ctx)
 		require.NoError(t, err)
 		assert.Equal(t, int64(step), int64(wait), "step #%d", i)
 	}