+++ timeutil/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+* Add clockctx package to enable a fake clock to be passed through a context.
+
 ## [0.2.0] - 2021-01-06
 
 ### Added